+++ examples/mcp-server/CHANGELOG.md
@@ -1,5 +1,43 @@
 # mcp-server-example
 
+## 0.0.1-alpha.32
+
+### Patch Changes
+
+- Updated dependencies [[`1fab6ca`](https://github.com/mattppal/mcut/commit/1fab6cada1e1310f196b7f05e49cf15145cad849)]:
+  - @mcut/timeline@0.1.0-alpha.19
+  - @mcut/mcp-server@0.1.0-alpha.33
+
+## 0.0.1-alpha.31
+
+### Patch Changes
+
+- Updated dependencies [[`ec1f719`](https://github.com/mattppal/mcut/commit/ec1f719b9a09f82dd95d548ec4a44f989895e8de)]:
+  - @mcut/timeline@0.1.0-alpha.18
+  - @mcut/mcp-server@0.1.0-alpha.32
+
+## 0.0.1-alpha.30
+
+### Patch Changes
+
+- Updated dependencies [[`b57daea`](https://github.com/mattppal/mcut/commit/b57daea9c55e044f93a00d90c5f18cf2c86365b5)]:
+  - @mcut/mcp-server@0.1.0-alpha.31
+
+## 0.0.1-alpha.29
+
+### Patch Changes
+
+- Updated dependencies [[`6827027`](https://github.com/mattppal/mcut/commit/682702705212c4cdd007e59c6a9ae2ec2ecd7f83)]:
+  - @mcut/mcp-server@0.1.0-alpha.30
+
+## 0.0.1-alpha.28
+
+### Patch Changes
+
+- Updated dependencies [[`74b5c22`](https://github.com/mattppal/mcut/commit/74b5c22a2c48244113e77ca5f3e03a1c23b80774)]:
+  - @mcut/mcp-server@0.1.0-alpha.29
+  - @mcut/timeline@0.1.0-alpha.17
+
 ## 0.0.1-alpha.27
 
 ### Patch Changes